+++ frontend/package.json
@@ -9,6 +9,7 @@
     "@material-ui/icons": "^4.11.3",
     "@material-ui/lab": "^4.0.0-alpha.57",
     "@mui/material": "^5.8.5",
+    "@mui/styles": "^5.8.6",
     "@testing-library/jest-dom": "^5.12.0",
     "@testing-library/react": "^11.2.5",
     "@testing-library/user-event": "^13.1.5",
