+++ frontend/src/components/CalcSpectrumPlot.tsx
