+++ frontend/src/components/fields/Mode.tsx
@@ -1,5 +1,12 @@
-import { FormControl, InputLabel, MenuItem, Select } from "@material-ui/core";
 import React from "react";
+// =================================================================================
+//                             MUI^5 module imports
+// ==================================================================================
+import FormControl from "@mui/material/FormControl";
+import InputLabel from "@mui/material/InputLabel";
+import MenuItem from "@mui/material/MenuItem";
+import Select from "@mui/material/Select";
+
 import { CalcSpectrumParams } from "../../constants";
 
 interface ModeProps {
