+++ frontend/src/components/fields/Pressure.tsx
@@ -1,5 +1,10 @@
-import { TextField, InputAdornment } from "@material-ui/core";
 import React from "react";
+// =================================================================================
+//                             MUI^5 module imports
+// ==================================================================================
+import TextField from "@mui/material/TextField";
+import InputAdornment from "@mui/material/InputAdornment";
+
 import { CalcSpectrumParams, ValidationErrors } from "../../constants";
 
 interface PressureProps {
