+++ frontend/src/components/fields/TGas.tsx
@@ -1,5 +1,11 @@
-import { InputAdornment, TextField } from "@material-ui/core";
+// importing react
 import React from "react";
+// =================================================================================
+//                             MUI^5 module imports
+// ==================================================================================
+import TextField from "@mui/material/TextField";
+import InputAdornment from "@mui/material/InputAdornment";
+
 import { CalcSpectrumParams, ValidationErrors } from "../../constants";
 
 interface TGasProps {
