+++ frontend/src/components/fields/WavenumberRangeSlider.tsx
@@ -1,9 +1,14 @@
 import React, { useState, useEffect } from "react";
-import { makeStyles } from "@material-ui/core/styles";
-import Grid from "@material-ui/core/Grid";
-import Typography from "@material-ui/core/Typography";
-import Slider from "@material-ui/core/Slider";
-import Input from "@material-ui/core/Input";
+
+// =================================================================================
+//                             MUI^5 module imports
+// ==================================================================================
+import { makeStyles } from "@mui/styles";
+import Grid from "@mui/material/Grid";
+import Typography from "@mui/material/Typography";
+import Slider from "@mui/material/Slider";
+import Input from "@mui/material/Input";
+
 import { CalcSpectrumParams } from "../../constants";
 
 interface WavelengthRangeSliderProps {
