--- website/src/radis.d.ts
@@ -1,9 +1,0 @@
-// declare module "*.jpg" {
-//   const value: string;
-//   export = value;
-// }
-// declare module "*.png" {
-//   const value: string;
-//   export = value;
-// }
-declare module "*.png";
